requirements.txt
google-cloud-firestore==2.3.4
pandas==1.3.5
streamlit==1.3.0
torch==1.10.1
transformers==4.15.0

